# openapi/v1/baselines/1.0.0.openapi.json
{
  "components": {
    "schemas": {
...
      "AdminOverviewResponseDto": {
        "properties": {
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/PermissionResponseDto"
            },
            "type": "array"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/AdminRoleResponseDto"
            },
            "type": "array"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/AdminUserResponseDto"
            },
            "type": "array"
          }
        },
        "required": [
          "users",
          "roles",
          "permissions"
        ],
        "type": "object"
      },
      "AdminRoleResponseDto": {
        "properties": {
          "active": {
            "type": "boolean"
          },
          "builtin": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": [
                "system.manage",
                "tasks.view",
                "tasks.create",
                "tasks.accept",
                "tasks.complete",
                "tasks.review",
                "tasks.close",
                "demo.reset"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "builtin",
          "permissions",
          "active",
          "deletedAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "AdminUserResponseDto": {
        "properties": {
          "active": {
            "type": "boolean"
          },
          "deletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "roleCode": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          },
          "roleName": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "name",
          "roleId",
          "roleCode",
          "roleName",
          "active",
          "deletedAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "ApiErrorBodyDto": {
        "properties": {
          "code": {
            "example": "CONFLICT",
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "message": {
            "example": "任务已被其他操作更新",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ApiErrorResponseDto": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBodyDto"
          },
          "path": {
            "example": "/api/v1/tasks/task-1/actions",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "error",
          "path",
          "timestamp"
        ],
        "type": "object"
      },
      "CreateAdminRoleDto": {
        "properties": {
          "name": {
            "example": "任务审核员",
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": [
                "system.manage",
                "tasks.view",
                "tasks.create",
                "tasks.accept",
                "tasks.complete",
                "tasks.review",
                "tasks.close",
                "demo.reset"
              ],
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "CreateAdminUserDto": {
        "properties": {
          "name": {
            "example": "新冒险家",
            "type": "string"
          },
          "roleId": {
            "example": "role-user",
            "type": "string"
          }
        },
        "required": [
          "name",
          "roleId"
        ],
        "type": "object"
      },
...
      "PermissionResponseDto": {
        "properties": {
          "code": {
            "enum": [
              "system.manage",
              "tasks.view",
              "tasks.create",
              "tasks.accept",
              "tasks.complete",
              "tasks.review",
              "tasks.close",
              "demo.reset"
            ],
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "description"
        ],
        "type": "object"
      },
...
      "UpdateAdminRoleDto": {
        "properties": {
          "name": {
            "example": "任务审核员",
            "type": "string"
          },
          "permissions": {
            "items": {
              "enum": [
                "system.manage",
                "tasks.view",
                "tasks.create",
                "tasks.accept",
                "tasks.complete",
                "tasks.review",
                "tasks.close",
                "demo.reset"
              ],
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "name",
          "permissions"
        ],
        "type": "object"
      },
      "UpdateAdminUserDto": {
        "properties": {
          "name": {
            "example": "新名字",
            "type": "string"
          },
          "roleId": {
            "example": "role-user",
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "demo-user": {
        "in": "header",
        "name": "X-Demo-User-Id",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "演示身份与任务状态机的 demo-only REST 契约",
    "title": "告示牌 Noticeboard API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/admin/overview": {
      "get": {
        "operationId": "getAdminOverview",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverviewResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/roles": {
      "post": {
        "operationId": "createAdminRole",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminRoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoleResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/roles/{id}": {
      "delete": {
        "operationId": "deleteAdminRole",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "updateAdminRole",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminRoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoleResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/roles/{id}/restore": {
      "post": {
        "operationId": "restoreAdminRole",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoleResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users": {
      "post": {
        "operationId": "createAdminUser",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}": {
      "delete": {
        "operationId": "deleteAdminUser",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "updateAdminUser",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{id}/restore": {
      "post": {
        "operationId": "restoreAdminUser",
        "parameters": [
          {
            "in": "header",
            "name": "X-Demo-User-Id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "demo-user": []
          }
        ],
        "summary": "",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "servers": [],
...
}